pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.1",
    "azure-storage-blob>=12.27.0",
    "fastapi[standard]>=0.120.0",
    "google-genai>=1.46.0",
    "langchain-google-genai>=4.2.1",
    "motor>=3.7.1",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "pymongo>=4.16.0",
]
